Rearrange and add some testing
